+++ cpp/src/umap/knn_graph/algo.cuh
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2024, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -126,11 +126,12 @@ inline void launcher(const raft::handle_t& handle,
 
     RAFT_EXPECTS(graph.distances().has_value(),
                  "return_distances for nn descent should be set to true to be used for UMAP");
-    auto out_knn_dists_view = raft::make_device_matrix_view(out.knn_dists, inputsA.n, n_neighbors);
+    auto out_knn_dists_view =
+      raft::make_device_matrix_view(out.knn_dists, inputsA.n, (uint64_t)n_neighbors);
     raft::matrix::slice<float, int64_t, raft::row_major>(
       handle, raft::make_const_mdspan(graph.distances().value()), out_knn_dists_view, coords);
     auto out_knn_indices_view =
-      raft::make_device_matrix_view(out.knn_indices, inputsA.n, n_neighbors);
+      raft::make_device_matrix_view(out.knn_indices, inputsA.n, (uint64_t)n_neighbors);
     raft::matrix::slice<int64_t, int64_t, raft::row_major>(
       handle, raft::make_const_mdspan(indices_d.view()), out_knn_indices_view, coords);
   }
